[feature/static]: [SSR] [Fix] don't fail if SQL file does not exist

+++ src/service/cli/generate-database-mock/fs-functions/trunctate-file.ts
@@ -4,8 +4,13 @@ import chalk from "chalk";
 
 async function truncateFile(filePath: string): Promise<void> {
   try {
-    await promises.truncate(filePath);
-    console.log(chalk.white(`File cleared: ${filePath}`));
+    const isFileExists = await promises.access(filePath).then(() => true).catch(() => false);
+    if (isFileExists) {
+      await promises.truncate(filePath);
+      console.log(chalk.white(`File cleared: ${filePath}`));
+    } else {
+      console.log(chalk.white(`File does not exist: ${filePath}`));
+    }
   } catch (e) {
     console.error(chalk.red(`Unable to clear file ${filePath}`));
     console.error(chalk.red(e));
